Rename trait_impls to type_impls
