+++ fediproto-sync/src/db/type_impls.rs
